Move login response to responses
