+++ lib/audiobookshelf_api.dart
@@ -1,6 +1,5 @@
 export 'src/audiobook_search_result.dart';
 export 'src/audiobookshelf_api_base.dart';
-export 'src/login_response.dart';
 export 'src/models/enums/file_type.dart';
 export 'src/models/enums/library_icon.dart';
 export 'src/models/enums/media_type.dart';
@@ -8,6 +7,7 @@ export 'src/models/enums/metadata_provider.dart';
 export 'src/models/enums/play_method.dart';
 export 'src/models/enums/schema_variant.dart';
 export 'src/models/enums/user_type.dart';
+export 'src/models/responses/login_response.dart';
 export 'src/models/schemas/audio_bookmark.dart';
 export 'src/models/schemas/audio_file.dart';
 export 'src/models/schemas/audio_meta_tags.dart';
